requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.0
gunicorn==20.1.0
h5py==3.1.0
json5==0.9.5
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
nltk==3.6.2
numpy==1.19.5
pickleshare==0.7.5
scikit-learn==0.24.2
tensorflow-cpu==2.5.0
